fix(framework): modify swagger
closes #2735
