+++ models/blueprint.go
@@ -29,12 +29,14 @@ const (
 	BLUEPRINT_MODE_ADVANCED = "ADVANCED"
 )
 
+// @Description CronConfig
 type Blueprint struct {
-	Name         string          `json:"name" validate:"required"`
-	Mode         string          `json:"mode" gorm:"varchar(20)" validate:"required,oneof=NORMAL ADVANCED"`
-	Plan         json.RawMessage `json:"plan"`
-	Enable       bool            `json:"enable"`
-	CronConfig   string          `json:"cronConfig" format:"* * * * *" example:"0 0 * * 1; please check https://crontab.guru/ for detail"`
+	Name   string          `json:"name" validate:"required"`
+	Mode   string          `json:"mode" gorm:"varchar(20)" validate:"required,oneof=NORMAL ADVANCED"`
+	Plan   json.RawMessage `json:"plan"`
+	Enable bool            `json:"enable"`
+	//please check this https://crontab.guru/ for detail
+	CronConfig   string          `json:"cronConfig" format:"* * * * *" example:"0 0 * * 1"`
 	IsManual     bool            `json:"isManual"`
 	Settings     json.RawMessage `json:"settings" swaggertype:"array,string" example:"please check api: /blueprints/<PLUGIN_NAME>/blueprint-setting"`
 	common.Model `swaggerignore:"true"`
