+++ plugins/tapd/api/connection.go
@@ -166,53 +166,3 @@ func GetConnection(input *core.ApiResourceInput) (*core.ApiResourceOutput, error
 	err := connectionHelper.First(connection, input.Params)
 	return &core.ApiResourceOutput{Body: connection}, err
 }
-
-// @Summary blueprints setting for tapd
-// @Description blueprint setting for tapd
-// @Tags plugins/tapd
-// @Accept application/json
-// @Param blueprint body TapdBlueprintSetting true "json"
-// @Router /blueprints/tapd/blueprint-setting [post]
-func PostTapdBlueprintSetting(input *core.ApiResourceInput) (*core.ApiResourceOutput, error) {
-	blueprint := &TapdBlueprintSetting{}
-	return &core.ApiResourceOutput{Body: blueprint, Status: http.StatusOK}, nil
-}
-
-type TapdBlueprintSetting []struct {
-	Version     string `json:"version"`
-	Connections []struct {
-		Plugin       string `json:"plugin"`
-		ConnectionID int    `json:"connectionId"`
-		Scope        []struct {
-			Options struct {
-				WorkspaceId uint64   `mapstruct:"workspaceId"`
-				CompanyId   uint64   `mapstruct:"companyId"`
-				Tasks       []string `mapstruct:"tasks,omitempty"`
-				Since       string
-			} `json:"options"`
-			Entities []string `json:"entities"`
-		} `json:"scope"`
-	} `json:"connections"`
-}
-
-// @Summary pipelines plan for tapd
-// @Description pipelines plan for tapd
-// @Tags plugins/tapd
-// @Accept application/json
-// @Param blueprint body TapdPipelinePlan true "json"
-// @Router /pipelines/tapd/pipeline-plan [post]
-func PostTapdPipelinePlan(input *core.ApiResourceInput) (*core.ApiResourceOutput, error) {
-	blueprint := &TapdPipelinePlan{}
-	return &core.ApiResourceOutput{Body: blueprint, Status: http.StatusOK}, nil
-}
-
-type TapdPipelinePlan [][]struct {
-	Plugin   string   `json:"plugin"`
-	Subtasks []string `json:"subtasks"`
-	Options  struct {
-		WorkspaceId uint64   `mapstruct:"workspaceId"`
-		CompanyId   uint64   `mapstruct:"companyId"`
-		Tasks       []string `mapstruct:"tasks,omitempty"`
-		Since       string
-	} `json:"options"`
-}
